fix(mcp): adapt to McpDisplayUrl and pub(crate) visibility after rebase
Upstream PRs #436/#438 made McpClientError pub(crate) with a private
McpDisplayUrl wrapper. Adapt call_tool to use resolved.display_url,
remove Connection::source field, and revert all mcp_client functions
and module to pub(crate) to match main. Move MCP client integration
tests back into the apis crate where they can access pub(crate) items.
