+++ apis/src/lib.rs
@@ -11,7 +11,7 @@
 
 pub mod anthropic;
 pub mod classifier;
-pub mod mcp_client;
+pub(crate) mod mcp_client;
 pub mod openai;
 #[cfg(feature = "store")]
 pub mod store;
