+++ tests/integration/tests/suite/main.rs
@@ -46,7 +46,6 @@ mod failure_mode;
 mod guardrails;
 mod mcp;
 mod mcp_broker;
-mod mcp_client;
 mod openai_mcp_tool_resolve;
 mod openai_responses_format;
 mod openai_responses_model_rewrite;
